:poop: Update Task2-3 in DAY3
않 이 진 짜 모 르 겠 어 으 어 어 으 어 어

언제 딱 한 번은 overflow가 발생 되었는데, 또 뭔가를 수정하니까 overflow 발생이 안 되는 것 같음.

+++ STM32/DAY3/Task2/Core/Src/main.c
@@ -139,7 +139,8 @@ int main(void)
 
 	TIM1->CCR1 = resultMotorSpeed;
 
-	encoder_value = (int16_t)TIM3->CNT;
+//	encoder_value = (int16_t)TIM3->CNT;
+	encoder_value = TIM3->CNT;
 	HAL_Delay(1); /* Debugging: Checking for the encoder_value */
   }
   /* USER CODE END 3 */
